+++ source/standalone/demo/play_cloner.py
@@ -142,7 +142,7 @@ def main():
         robot.set_joint_position_target(actions)
         robot.write_data_to_sim()
         # perform step
-        sim.step(render=app_launcher.RENDER)
+        sim.step()
         # update sim-time
         sim_time += sim_dt
         ep_step_count += 1